`dotnet watch run` shouldn't disable the browser refresh logic injection on Linux (#15340)
Prior to this change, inability or failure to launch a browser would result in the browser-refresh feature from being disabled.
On Linux the feature is a bit flaky so we always disabled it. This change loosens up some of the restrictions.
We always attempt to launch a browser and use heuristics to determine if launching failed. Additionally,
the browser refresh feature is enabled regardless of how launching the browser fared.

Fixes #27081