Use ProjectCapability to determine hot reloadibility (#17384)
* Use ProjectCapability to determine hot reloadibility

dotnet-watch currently relies on the presence of a value in the launchSettings.json to
determine if a project supports hot reload. This is problematic for workloads other than
ASP.NET Core that do not contain a launchSettings.

This change introduces a ProjectCapability in the SDK that is configured for projects targeting 6.0 or newer
and updates dotnet-watch to start using this capability. In addition, for the "esoteric" cases of BlazorWasm
and BlazorWasm-hosted, we'll rely on the presence of additional capabilities rather than the profile.