Correctly initialize RefreshBrowserAfterFileChange (#18426)
dotnet-watch uses a heuristic to determine if it should rely on the app
refreshing vs performing a browser refresh. It decides it should not refresh
the browser for any Blazor app (app with Blazor components). This heuristic
was not being correctly initialized after a rude edit.

Fixes https://github.com/dotnet/aspnetcore/issues/33655